+++ lib/flutter_password_input.dart
@@ -1,4 +1,5 @@
-export 'package:just_tooltip/just_tooltip.dart' show TooltipAnimation;
+export 'package:just_tooltip/just_tooltip.dart'
+    show TooltipAlignment, TooltipAnimation, TooltipDirection;
 
 export 'src/password_text_field.dart';
 export 'src/password_text_field_theme.dart';